+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ### Updated
 - `@reportportal/client-javascript` bumped to version `5.0.14`
 ### Added
-- Browser parameter to steps
+- Browser parameter to steps.
+### Changed
+- `token` configuration option was renamed to `apiKey` to maintain common convention.
+- `@reportportal/client-javascript` bumped to version `5.0.12`.
+- Readme file updated.
 
 ## [5.0.3] - 2022-10-05
 ### Added