+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+
+## [5.1.0] - 2024-02-06
 ### Added
 - `launchId` option to the config to attach run results to an existing launch. Related to parallel execution on one and several machines.
 - Browser parameter to steps.